Rework and restructure typing
