+++ src/index.d.ts
@@ -1,13 +1,13 @@
 type TokenType = 'Punctuation' | 'Note' | 'FunctionWord' | 'Word' | 'Clause' | 'Added' | 'Phrase'
 
-interface MessagedToken {
+type TokenBase = {
 	token: string
 	type: TokenType
 	tag: Tag
 	messages: Message[]
 }
 
-interface Token extends MessagedToken {
+type Token = TokenBase & {
 	specified_sense: string
 	lookup_terms: LookupTerm[]
 	lookup_results: LookupResult[]
@@ -28,12 +28,12 @@ type Sentence = {
 // Messages
 type MessageLabel = 'error' | 'warning' | 'suggest' | 'info'
 
-interface MessageType {
+type MessageType = {
 	label: MessageLabel
 	severity: number
 }
 
-interface Message extends MessageType {
+type Message = MessageType & {
 	message: string
 	// TODO #101 add optional fix_action structure
 }
@@ -48,12 +48,12 @@ type MessageInfo = {
 // Lookup values
 type LookupTerm = string
 
-interface LookupWord {
+type LookupWord = {
 	stem: string
 	part_of_speech: string
 }
 
-interface LookupResult extends LookupWord {
+type LookupResult = LookupWord & {
 	lexicon_id: number
 	form: string
 	// TODO include features
@@ -70,62 +70,4 @@ type HowToEntry = {
 	structure: string
 	pairing: string
 	explication: string
-}
-
-// External API Responses
-interface LexicalFormResult extends LookupWord {
-	id: number
-	form: string
-}
-
-interface OntologyResult extends LookupWord {
-	id: string
-	sense: string
-	level: number
-	gloss: string
-	categorization: string
-}
-
-interface HowToResult extends HowToEntry {
-	term: string
-	part_of_speech: string
-}
-
-// Internal API response
-type CheckResponse = {
-	has_error: boolean
-	tokens: SimpleToken[]
-	back_translation: string
-}
-
-interface SimpleToken extends MessagedToken {
-	lookup_results: SimpleLookupResult[]
-	complex_pairing: SimpleToken | null
-	pronoun: SimpleToken | null
-	sub_tokens: SimpleToken[]
-}
-
-interface SimpleLookupResult extends LookupWord {
-	lexicon_id: number
-	form: string
-	// TODO include features
-	ontology_id: number
-	sense: string
-	level: number
-	gloss: string
-	categorization: string
-	how_to_entries: HowToEntry[]
-	case_frame: SimpleCaseFrameResult
-}
-
-type SimpleCaseFrameResult = {
-	is_valid: boolean
-	is_checked: boolean
-	valid_arguments: SimpleRoleArgResult
-	extra_arguments: SimpleRoleArgResult
-	missing_arguments: RoleTag[]
-}
-
-type SimpleRoleArgResult = {
-	[role: RoleTag]: string
 }
