+++ src/lib/lookups/ontology.js
@@ -1,6 +1,6 @@
 import { PUBLIC_ONTOLOGY_API_HOST } from '$env/static/public'
 import { LOOKUP_FILTERS } from '$lib/lookup_filters'
-import { create_lookup_result } from '$lib/parser/token'
+import { create_lookup_result } from '$lib/token'
 
 /**
  * @param {Token} lookup_token
