+++ src/lib/parser/clausify.js
@@ -1,6 +1,6 @@
 import { REGEXES } from '$lib/regexes'
 import { ERRORS } from './error_messages'
-import { TOKEN_TYPE, create_clause_token, create_added_token, flatten_sentence, MESSAGE_TYPE } from './token'
+import { TOKEN_TYPE, create_clause_token, create_added_token, flatten_sentence, MESSAGE_TYPE } from '../token'
 
 /**
  * @param {Token[]} tokens
