+++ src/lib/rules/case_frame/verbs/presets.js
@@ -1,4 +1,4 @@
-import { TOKEN_TYPE } from "$lib/parser/token"
+import { TOKEN_TYPE } from '$lib/token'
 
 /**
  * 
@@ -104,7 +104,7 @@ export function predicate_adjective() {
 			'category': 'Adjective',
 			'tag': { 'adj_usage': 'predicative' },
 			'skip': ['adjp_modifiers_predicative', 'advp', { 'tag': 'verb_polarity' }],
-		}
+		},
 	})
 }
