+++ src/lib/rules/part_of_speech_rules.js
@@ -1,6 +1,6 @@
 import { ERRORS } from '$lib/parser/error_messages'
 import { create_context_filter, create_token_filter, message_set_action, simple_rule_action } from './rules_parser'
-import { TOKEN_TYPE, create_lookup_result, is_one_part_of_speech } from '$lib/parser/token'
+import { TOKEN_TYPE, create_lookup_result, is_one_part_of_speech } from '$lib/token'
 import { LOOKUP_FILTERS } from '$lib/lookup_filters'
 
 /**
